Move the db module to storage
The `db` is a part of the storage layer (along with the `seed`
module). We can make this relationship explicit by putting the `db`
module under the `storage` module.
