+++ cnd/src/proptest.rs
@@ -211,13 +211,13 @@ pub mod hbit {
 
 pub mod db {
     use super::*;
-    use crate::db;
+    use crate::storage;
     use std::fmt::Debug;
 
     pub fn created_swap<A, B>(
         alpha: impl Strategy<Value = A>,
         beta: impl Strategy<Value = B>,
-    ) -> impl Strategy<Value = db::CreatedSwap<A, B>>
+    ) -> impl Strategy<Value = storage::CreatedSwap<A, B>>
     where
         A: Debug,
         B: Debug,
@@ -231,7 +231,7 @@ pub mod db {
             timestamp(),
         )
             .prop_map(|(swap_id, alpha, beta, peer, role, start_of_swap)| {
-                db::CreatedSwap {
+                storage::CreatedSwap {
                     swap_id,
                     alpha,
                     beta,
@@ -245,7 +245,7 @@ pub mod db {
 
     pub mod tables {
         use super::*;
-        use db::{tables, tables::IntoInsertable};
+        use storage::{tables, tables::IntoInsertable};
 
         prop_compose! {
             pub fn insertable_swap()(
